requirements.txt
PyGithub
ruamel.yaml
